Refactored reducers so the structure isn't so ridiculous now
