+++ imports/ui/actions/favourites.js
@@ -5,4 +5,3 @@ export const toggleFavourites = () => {
 		type: TOGGLE_FAVOURITES
 	}	
 }
-
