+++ imports/ui/actions/input.js
@@ -1,0 +1,5 @@
+export const UPDATE_INPUT = 'UPDATE_INPUT';
+
+export const updateInput = val => {
+	return { type: UPDATE_INPUT, payload: val }
+}
