+++ imports/ui/actions/recipe.js
@@ -1,0 +1,8 @@
+export const SET_RECIPE_DETAILS = 'SET_RECIPE_DETAILS';
+
+export const setRecipeDetails = recipe => {
+  return {
+    type: SET_RECIPE_DETAILS,
+    recipe: recipe
+  }
+}
