+++ imports/ui/components/LoginButton.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Blaze from 'meteor/gadicc:blaze-react-component';
 
-
 class LoginButton extends React.Component {
 	constructor(props) {
 		super(props);
