+++ imports/ui/components/RecipeCards.jsx
@@ -9,10 +9,11 @@ import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import CardMedia from "@material-ui/core/CardMedia";
 import Button from "@material-ui/core/Button";
-import { setRecipeDetails, openDetailedView, closeDetailedView } from '../actions';
+import { openDetailedView, closeDetailedView } from '../actions/detailedView.js';
+import { setRecipeDetails } from '../actions/recipe.js';
 import RecipeDetails from './RecipeDetails';
 import '../style/RecipeCards.css';
-import Favourites from '../../api/favourites'
+import Favourites from '../../api/favourites';
 import Icon from '@material-ui/core/Icon';
 
 class RecipeCards extends Component {
