+++ imports/ui/reducers/recipeDetailsReducer.js
@@ -1,0 +1,10 @@
+import { SET_RECIPE_DETAILS } from '../actions/recipe.js';
+import { EMPTY_RECIPE } from '../../util/Recipe.jsx';
+
+export const recipeDetailsReducer = (recipe = EMPTY_RECIPE, action) => {
+	if (action.type === SET_RECIPE_DETAILS) {
+		return action.recipe;
+	} else {
+    return recipe;
+  }
+};
